refactoring tests

+++ src/test/java/api/BaseTest.java
@@ -13,6 +13,7 @@ import org.assertj.core.api.SoftAssertions;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.TestInfo;
 import org.junit.jupiter.api.extension.ExtendWith;
 
 
@@ -36,9 +37,11 @@ public class BaseTest {
     }
 
     @AfterEach
-    public void afterTest(){
+    public void afterTest(TestInfo testInfo){
         softly.assertAll();
-        SessionStorage.deleteAllViaApi();
+        if (!testInfo.getDisplayName().equals("adminCanDeletePerson()")) {
+            SessionStorage.deleteAllViaApi();
+        }
         SessionStorage.clear();
     }
 }
